FriendTech_Discord_Integration/requirements.txt
aiohttp==3.8.5
discord.py==2.3.2
python-dotenv==1.0.0
python-json-logger==2.0.7
websockets==11.0.3
